Refactor: calculateTimeAgo 함수 분리
